requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
reportlab==4.0.9
cryptography==41.0.7
